+++ database/upgrade/upgrade4.6.0RC.sql
@@ -1,0 +1,21 @@
+-- ---------------------------------------------------------------------------------------------------------------------
+-- WARNING: BACKUP YOUR DATABASE BEFORE UPGRADING
+-- THIS SCRIPT IS ONLY FOR UPGRADING 4.5.0 TO 4.6.0
+-- THE CURRENT VERSION CAN BE FOUND AT `myems_system_db`.`tbl_versions`
+-- ---------------------------------------------------------------------------------------------------------------------
+
+START TRANSACTION;
+
+CREATE TABLE IF NOT EXISTS `myems_historical_db`.`tbl_energy_plan_files` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `file_name` VARCHAR(255) NOT NULL,
+  `uuid` CHAR(36) NOT NULL,
+  `upload_datetime_utc` DATETIME NOT NULL,
+  `status` VARCHAR(45) NOT NULL COMMENT 'new, done, error',
+  `file_object` LONGBLOB NOT NULL,
+  PRIMARY KEY (`id`));
+
+-- UPDATE VERSION NUMBER
+UPDATE `myems_system_db`.`tbl_versions` SET version='4.6.0RC', release_date='2024-06-18' WHERE id=1;
+
+COMMIT;